feat(systemd): add helper for readiness process child_spec
This function is meant to be used as part of the supervisor child list
in the place where system is ready (there still can be processes to be
started, but these should not be vital to the whole system). It will
just start temporary process that send message and exits.