feat(systemd_kmsg_formatter): automatically set kmsg formatter
Now the `systemd` application will check the device and inode number set
to the `stdout` and/or `stderr` on startup and will automatically set
proper formatting to include the level prefix to all lines of the
output to provide information for journal. This can be disabled by
setting `systemd.auto_kmsg` environment option to `false` (discouraged).