fix: Add Erlang 24 support (#29)
* GitHub Actions: Add Erlang 24 to "erlang" workflow

* GitHub Actions: Turn off fail-fast for testing on Erlang 23/24

* *_SUITE: Adapt to new return value from `socket:bind/2`

`socket:bind/2` returned an `{ok, _}` tuple up to and including Erlang
23. Starting with Erlang 24, it returns `ok`. If the call fails, it will
trigger a case clause exception in both cases.

* systemd_test_utils: Convert `Path` from binary to string if necessary

`socket:sockname/1` seems to return a binary in Erlang 24, not a string.
This must be converted to a string before passing it to `os:putenv/2`.

* systemd_protocol: Encode message as an iovec()

`socket:sendmsg/2` expects an iovec() in the `iov` message map entry. An
iovec() is a flat list of binaries.

The systemd_protocol module is modified to product just that:
    * Strings and characters (integers) are converted to binaries.
    * The final list is flatten, i.e. the nested lists of binaries are
      converted to a list with binaries at the same level.

This fixes the library with Erlang 24. Before this patch,
`socket:sendmsg/2` would return the following error:

    {error, {invalid, {iov, {element_not_binary, 1}}}}

Co-authored-by: Philip Kuryloski <kuryloskip@vmware.com>