ft: add warning log message if there is no readiness message sent
This is one of the most common issues that people encounter from my
experience when they are using this library together with `Type=notify`
systemd units. This adds message to warn user that they may have forgot
about sending that message. This should help with debugging reason for
seemingly unreasonable application restarts.