ft: helper function for accessing credentials directory
This adds new function `systemd:credentials/1` which returns path for
credentials' files. This function checks if the `$CREDENTIALS_DIRECTORY`
variable was set and if it was set, then the path is always returned. It
is left up to the user to check if the file is existing and/or correct
one.

Close #41